Updated Reshape operator with a default value. (#317)
* Fix type

* Avoid mmap error

* Use QUANTIZED_PACKED

* Updated Reshape operator with a default value.

dimension_format for Reshape had no default value and there were several
places which used it without providing one. I opted for setting
a default value rather than updating all places to supply
one of their own. Not sure why something like Reshape needs
this parameter though; seems like the superclass requires it so
we have it in Reshape as a consequence. If that is the only reason,
removing it from the Reshape constructor and using the
superclass's default would be better.