Add turbo color map (#1055)
## What this patch does to fix the issue.

This PR contains replacement of the current `Jet` colormap of matplotlib by adding the Google proposed [Turbo Colormap](https://ai.googleblog.com/2019/08/turbo-improved-rainbow-colormap-for.html) as highlighted in #485 by @tk26eng in `blueoil/common.py`. 

This colormap is for replacing the `Jet` colormap in matplotlib that is used for coloring heatmaps. So I added a new function that takes the heatmap as input and then returns the colored heatmap. As the function uses numpy, it should be called using `tf.py_func` interface.

The table for reference colors is a (256x3) matrix of sRGB colors and I saved it to a separate file `blueoil/turbo_color_map.py`.

## Link to any relevant issues or pull requests.

Relevant issue: #485